feat: created table users on page users

+++ server/controllers/user.js
@@ -1,5 +1,8 @@
 const db = require('../db');
-const userValidation = require('../validations/user');
+const [
+  CreateUserValidation,
+  UpdateUserValidation,
+] = require('../validations/user');
 const bcrypt = require('bcrypt');
 class UserController {
   async getUsers(req, res) {
@@ -54,7 +57,7 @@ class UserController {
   }
 
   async createUser(req, res) {
-    const { error } = userValidation(req.body);
+    const { error } = CreateUserValidation(req.body);
     if (error) {
       const errorMessage = error.details[0].message;
       return res.status(400).json({ error: errorMessage });
@@ -72,7 +75,11 @@ class UserController {
         role,
         avatar,
       } = req.body;
-      password = await bcrypt.hash(password, 3);
+      if (phone && phone.startsWith('0')) {
+        phone = '+38' + phone;
+      } else if (phone && phone.startsWith('380')) {
+        phone = '+' + phone;
+      }
       const phoneAlreadyExists = await db.query(
         'SELECT * FROM users WHERE phone = $1',
         [phone]
@@ -89,6 +96,7 @@ class UserController {
       if (emailAlreadyExists.rows[0]) {
         return res.status(409).json('Користувач з таким email вже існує');
       }
+      password = await bcrypt.hash(password, 3);
       if (!gender) {
         gender = 'man';
       }
@@ -115,7 +123,7 @@ class UserController {
     } catch (error) {
       return res
         .status(500)
-        .json('Не вдалось виконати запит, спробуйте пізніше');
+        .json(error.message);
     }
 
     // #swagger.tags = ['Users']
@@ -160,10 +168,10 @@ class UserController {
     try {
       const id = req.params.id;
       const user = await db.query('SELECT * FROM users WHERE id = $1', [id]);
-      if (!user.rows[0]) {
-        return res.status(400).json('Такого користувача не існує');
-      } else {
+      if (user.rows[0]) {
         res.json(user.rows[0]);
+      } else {
+        return res.status(400).json('Такого користувача не існує');
       }
     } catch (error) {
       return res
@@ -194,10 +202,12 @@ class UserController {
   }
 
   async updateUser(req, res) {
-    const { error } = userValidation(req.body);
+    const { error } = UpdateUserValidation(req.body);
     if (error) {
+      console.log('error', error);
       const errorMessage = error.details[0].message;
-      return res.status(400).json({ error: errorMessage });
+      console.log('errorMessage', errorMessage);
+      return res.status(400).json(errorMessage);
     }
     try {
       let {
@@ -213,7 +223,13 @@ class UserController {
         role,
         avatar,
       } = req.body;
-      password = await bcrypt.hash(password, 3);
+      console.log('req.body',req.body);
+      if (phone && phone.startsWith('380')) {
+        phone = '+' + phone;
+      }
+      if (phone && phone.startsWith('0')) {
+        phone = '+38' + phone;
+      }
       const phoneAlreadyExists = await db.query(
         'SELECT * FROM users WHERE phone = $1',
         [phone]
@@ -234,38 +250,63 @@ class UserController {
           return res.status(409).json('Користувач з таким email вже існує');
         }
       }
-      if (!gender) {
-        gender = 'man';
+      let updateFields = {};
+      if (name) {
+        updateFields.name = name;
       }
-      if (!role) {
-        role = 'user';
+      if (surname) {
+        updateFields.surname = surname;
       }
-      const user = await db.query(
-        'UPDATE users set name = $2, surname = $3, gender=$4, phone = $5, password = $6,' +
-          'email = $7, floor = $8, room = $9, role = $10, avatar = $11 WHERE id = $1 RETURNING *',
-        [
-          id,
-          name,
-          surname,
-          gender,
-          phone,
-          password,
-          email,
-          floor,
-          room,
-          role,
-          avatar,
-        ]
-      );
-      if (!user.rows[0]) {
+      if (gender) {
+        updateFields.gender = gender;
+      }
+      if (phone) {
+        updateFields.phone = phone;
+      }
+      if (password) {
+        updateFields.password = await bcrypt.hash(password, 3);
+      }
+      if (email) {
+        updateFields.email = email;
+      }
+      if (floor) {
+        updateFields.floor = floor;
+      }
+      if (room) {
+        updateFields.room = room;
+      }
+      if (role) {
+        updateFields.role = role;
+      }
+      if (avatar) {
+        updateFields.avatar = avatar;
+      }
+      console.log('updateFields', updateFields);
+      if (Object.keys(updateFields).length === 0) {
+        return res.status(400).json('Не вказано жодного поля для оновлення');
+      }
+      const updateUserQuery = {
+        text: `
+        UPDATE users
+        SET ${Object.keys(updateFields)
+          .map((key, index) => `${key} = $${index + 2}`)
+          .join(', ')}
+        WHERE id = $1
+        RETURNING *
+      `,
+        values: [id, ...Object.values(updateFields)],
+      };
+      const UpdateUser = await db.query(updateUserQuery);
+      if (!UpdateUser.rows[0]) {
         return res.status(400).json('Такого користувача не існує');
       } else {
-        res.json(user.rows[0]);
+        res.json(UpdateUser.rows[0]);
       }
     } catch (error) {
+      console.dir(error.message);
       return res
         .status(500)
-        .json('Не вдалось виконати запит, спробуйте пізніше');
+        .json(error.message);
     }
 
     // #swagger.tags = ['Users']
@@ -310,8 +351,13 @@ class UserController {
   async deleteUser(req, res) {
     try {
       const id = req.params.id;
-      await db.query('DELETE FROM users WHERE id = $1', [id]);
-      return res.status(200).json('Видалено');
+      const user = await db.query('SELECT * FROM users WHERE id = $1', [id]);
+      if (user.rows[0]) {
+        await db.query('DELETE FROM users WHERE id = $1', [id]);
+        return res.status(200).json('Видалено');
+      } else {
+        return res.status(400).json('Такого користувача не існує');
+      }
     } catch (error) {
       return res
         .status(500)
